requirements.txt
azure-common==1.1.23
azure-cosmos==3.1.2
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-functions==1.0.4
azure-nspkg==3.0.2
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
idna==2.8
pycparser==2.19
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
urllib3==1.25.6
yapf==0.28.0
